+++ inst/include/geojson_wkt.h
@@ -1,19 +1,11 @@
 
-#ifndef GEOJSON_WKT_H
-#define GEOJSON_WKT_H
+#ifndef GEOJSONSF_WKT_H
+#define GEOJSONSF_WKT_H
 
 #include "rapidjson/document.h"
 using namespace rapidjson;
 
 void coord_separator(std::ostringstream& os, int i, int n);
-
-void line_separator_wkt(std::ostringstream& os, int i, int n);
-
-void object_separator(std::ostringstream& os);
-
-void polygon_separate_wkt(std::ostringstream& os, int i, int n);
-
-void add_lonlat_to_wkt_stream(std::ostringstream& os, float lon, float lat );
 
 void begin_wkt(std::ostringstream& os, std::string& geom_type);