+++ src/geojson_sfc.cpp
@@ -2,8 +2,6 @@
 #include "geojsonsf.h"
 #include "geojson_sfc.h"
 
-using namespace Rcpp;
-
 void calculate_bbox(Rcpp::NumericVector& bbox, Rcpp::NumericVector& point) {
   //xmin, ymin, xmax, ymax
   bbox[0] = std::min(point[0], bbox[0]);
@@ -16,7 +14,7 @@ void calculate_bbox(Rcpp::NumericVector& bbox, Rcpp::NumericVector& point) {
 
 std::string attach_class(Rcpp::List& sfc,
                          std::string geom_type,
-                         std::set< std::string >& geometry_types) {
+                         std::unordered_set< std::string >& geometry_types) {
 
   std::string geometry_class;
   if (geom_type == "GEOMETRYCOLLECTION") {
@@ -48,7 +46,7 @@ std::string attach_class(Rcpp::List& sfc,
 void attach_sfc_attributes(Rcpp::List& sfc,
                            std::string& type,
                            Rcpp::NumericVector& bbox,
-                           std::set< std::string >& geometry_types,
+                           std::unordered_set< std::string >& geometry_types,
                            int& nempty) {
 
   std::string geometry_class = attach_class(sfc, type, geometry_types);
@@ -57,8 +55,10 @@ void attach_sfc_attributes(Rcpp::List& sfc,
   double prec = 0;
 
   // attribute::crs
-  Rcpp::List crs = Rcpp::List::create(Named("epsg") = geojsonsf::EPSG,
-                                     Named("proj4string") = geojsonsf::PROJ4STRING);
+  Rcpp::List crs = Rcpp::List::create(
+  	Rcpp::Named("epsg") = geojsonsf::EPSG,
+  	Rcpp::Named("proj4string") = geojsonsf::PROJ4STRING
+  	);
 
   crs.attr("class") = Rcpp::CharacterVector::create("crs");
   sfc.attr("crs") = crs;
@@ -81,12 +81,6 @@ Rcpp::NumericVector start_bbox() {
   return bbox;
 }
 
-std::set< std::string> start_geometry_types() {
-  std::set< std::string> geometry_types;
-  return geometry_types;
-}
-
-
 Rcpp::StringVector start_sfc_classes(size_t collectionCount) {
   Rcpp::StringVector sfc_classes(collectionCount);
   return sfc_classes;
@@ -101,7 +95,7 @@ void fetch_geometries(Rcpp::List& sf, Rcpp::List& res, int& sfg_counter) {
     switch( TYPEOF(*it) ) {
 
     case VECSXP: {
-      Rcpp::List tmp = as<Rcpp::List>(*it);
+      Rcpp::List tmp = Rcpp::as< Rcpp::List >( *it );
       if(Rf_isNull(tmp.attr("class"))){
         fetch_geometries(tmp, res, sfg_counter);
       } else {
@@ -111,8 +105,8 @@ void fetch_geometries(Rcpp::List& sf, Rcpp::List& res, int& sfg_counter) {
       break;
     }
     case REALSXP: {
-      Rcpp::NumericVector tmp = as<Rcpp::NumericVector>(*it);
-      if(Rf_isNull(tmp.attr("class"))){
+      Rcpp::NumericVector tmp = Rcpp::as< Rcpp::NumericVector >( *it );
+      if(Rf_isNull(tmp.attr("class"))) {
         Rcpp::stop("Geometry could not be determined");
       } else {
         res[sfg_counter] = tmp;
@@ -121,8 +115,8 @@ void fetch_geometries(Rcpp::List& sf, Rcpp::List& res, int& sfg_counter) {
       break;
     }
     case INTSXP: {
-      Rcpp::IntegerVector tmp = as<Rcpp::IntegerVector>(*it);
-      if(Rf_isNull(tmp.attr("class"))){
+      Rcpp::IntegerVector tmp = Rcpp::as< Rcpp::IntegerVector >( *it );
+      if(Rf_isNull( tmp.attr( "class" ) ) ){
         Rcpp::stop("Geometry could not be determined");
       } else {
         res[sfg_counter] = tmp;
@@ -131,8 +125,8 @@ void fetch_geometries(Rcpp::List& sf, Rcpp::List& res, int& sfg_counter) {
       break;
     }
     case STRSXP: {
-    	Rcpp::StringVector tmp = as<Rcpp::StringVector>(*it);
-    	if(Rf_isNull(tmp.attr("class"))){
+    	Rcpp::StringVector tmp = Rcpp::as< Rcpp::StringVector >( *it );
+    	if(Rf_isNull( tmp.attr( "class" ) ) ) {
     		Rcpp::stop("Geometry could not be determined");
     	} else {
     		res[sfg_counter] = tmp;
@@ -151,16 +145,16 @@ void fetch_geometries(Rcpp::List& sf, Rcpp::List& res, int& sfg_counter) {
 Rcpp::List construct_sfc(int& sfg_objects,
                          Rcpp::List& sf,
                          Rcpp::NumericVector& bbox,
-                         std::set< std::string >& geometry_types,
+                         std::unordered_set< std::string >& geometry_types,
                          int& nempty) {
 
-  Rcpp::List sfc_output(sfg_objects);
+  Rcpp::List sfc_output( sfg_objects );
   std::string geom_attr;
 
   int sfg_counter = 0;
 
-  fetch_geometries(sf, sfc_output, sfg_counter);
-  attach_sfc_attributes(sfc_output, geom_attr, bbox, geometry_types, nempty);
+  fetch_geometries( sf, sfc_output, sfg_counter );
+  attach_sfc_attributes( sfc_output, geom_attr, bbox, geometry_types, nempty );
 
   return sfc_output;
 }